修复音量函数的bug

+++ include/volume.h
@@ -7,4 +7,4 @@ typedef enum {
     AUDIO_VOLUME_GET,
     AUDIO_VOLUME_PLUS,
 } audio_volume_action;
-int audio_volume(audio_volume_action action, long *outvol);
+int audio_volume(audio_volume_action action, int *outvol);
